Converted to slider
Basic UI should be done.
What I did to do this, was create a slider, then within movement, I reference that slider by drag-in-dropping, setting the value directly, and then boom: done.

One thing to watch out for is how I implented the gunCooldown. Before, cooldown approches -∞, but here I set the cap to be 0 using a tetary operator.